realocate like and comment related enpoints to post. router
