+++ config/cloudinary.js
@@ -1,4 +1,3 @@
-require("dotenv").config();
 const cloudinary = require("cloudinary").v2;
 cloudinary.config({
   cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
