+++ routers/auth.js
@@ -54,6 +54,7 @@ router.post("/login", async (req, res, next) => {
 
 router.post("/signup", async (req, res) => {
   const { email, password, name } = req.body;
+  // console.log("req body", req.body);
   if (!email || !password || !name) {
     return res.status(400).send("Please provide an email, password and a name");
   }
@@ -68,7 +69,7 @@ router.post("/signup", async (req, res) => {
     delete newUser.dataValues["password"]; // don't send back the password hash
 
     const token = toJWT({ userId: newUser.id });
-
+    // console.log("new user", newUser);
     res.status(201).json({ token, ...newUser.dataValues });
   } catch (error) {
     if (error.name === "SequelizeUniqueConstraintError") {
@@ -97,15 +98,19 @@ router.get("/me", authMiddleware, async (req, res) => {
   delete req.user.dataValues["password"];
   res.status(200).send({ ...req.user.dataValues, posts, likes });
 });
+
 router.get("/:userId", async (req, res) => {
   const { userId } = req.params;
-  const posts = await Post.findAll({
+  const user = await User.findOne({
     where: { userId: userId },
-    include: { model: Image },
-    order: [["createdAt", "DESC"]],
+    include: {
+      model: Post,
+      include: { model: Image },
+      order: [["createdAt", "DESC"]],
+    },
   });
 
-  res.status(200).send(posts);
+  res.status(200).send(user);
 });
 
 // update personal infor: name, description
@@ -133,165 +138,4 @@ router.patch("/:userId/profilePic", authMiddleware, async (req, res) => {
   }
 });
 
-// like a post
-router.post("/post/:postId/like", authMiddleware, async (req, res) => {
-  try {
-    const { userId } = req.body;
-    const postId = req.params.postId;
-    const like = await Like.findOne({
-      where: { userId: userId, postId: postId },
-    });
-
-    const post = await Post.findOne({
-      where: { id: postId },
-      include: [
-        { model: Image },
-        { model: User, attributes: ["name", "profile_pic"] },
-        {
-          model: Comment,
-          include: [{ model: User, attributes: ["name", "profile_pic"] }],
-          order: [["createdAt", "DESC"]],
-        },
-      ],
-    });
-    if (post === null) {
-      res.status(400).send({ error: "Post does not exist" });
-    } else {
-      if (like === null) {
-        const newLike = await Like.create({
-          userId: userId,
-          postId: postId,
-        });
-        post.likes_num++;
-        await post.update({ likes_num: post.likes_num });
-        return res.status(200).send({ message: "Like successfully", post });
-      } else {
-        return res.status(400).send({ error: "You already liked this post" });
-      }
-    }
-  } catch (error) {
-    console.log(error);
-  }
-});
-
-router.get("/post/:postId/allLikes", authMiddleware, async (req, res) => {
-  const likes = await Like.findAll({ where: { postId: req.params.postId } });
-  res.status(200).send({ message: "All likes of this post", likes });
-});
-
-//unlike a post
-router.post("/post/:postId/unlike", authMiddleware, async (req, res) => {
-  try {
-    const { userId } = req.body;
-    const postId = req.params.postId;
-    const like = await Like.findOne({
-      where: { userId: userId, postId: postId },
-    });
-
-    const post = await Post.findOne({
-      where: { id: postId },
-      include: [
-        { model: Image },
-        { model: User, attributes: ["name", "profile_pic"] },
-        {
-          model: Comment,
-          include: [{ model: User, attributes: ["name", "profile_pic"] }],
-          order: [["createdAt", "DESC"]],
-        },
-      ],
-    });
-    if (post === null) {
-      res.status(400).send({ error: "Post does not exist" });
-    } else {
-      if (like === null) {
-        return res.status(400).send({ error: "You already unliked this post" });
-      } else {
-        await Like.destroy({ where: { userId: userId, postId: postId } });
-        post.likes_num--;
-        await post.update({ likes_num: post.likes_num });
-        return res.status(200).send({ message: "Unliked successfully", post });
-      }
-    }
-  } catch (error) {
-    console.log(error);
-  }
-});
-
-// add a new comment
-router.get("/post/:postId/allComments", async (req, res) => {
-  const comments = await Comment.findAll({
-    where: { postId: req.params.postId },
-  });
-  res.status(200).send({ message: "All comments of this post", comments });
-});
-
-router.post(
-  "/post/:postId/comment",
-  authMiddleware,
-  async (request, response) => {
-    try {
-      const user = await User.findByPk(request.user.id);
-      const postId = request.params.postId;
-
-      const { content } = request.body; // reveive data from request
-      if (!content) {
-        return response
-          .status(400)
-          .send({ message: "A comment must have a content" });
-      }
-
-      const newComment = await Comment.create({
-        content,
-        userId: user.id,
-        postId: postId,
-      });
-      const post = await Post.findOne({
-        where: { id: postId },
-        include: [
-          { model: Image },
-          { model: User, attributes: ["name", "profile_pic"] },
-          {
-            model: Comment,
-            include: [{ model: User, attributes: ["name", "profile_pic"] }],
-            order: [["createdAt", "DESC"]],
-          },
-        ],
-      });
-      await post.update({ comments_num: post.comments_num + 1 });
-      return response.status(201).send(post);
-    } catch (error) {
-      console.log(error);
-    }
-  }
-);
-
-// delete a comment
-router.delete(
-  "/post/:postId/comment/:commentId/delete",
-  authMiddleware,
-  async (request, response) => {
-    try {
-      const userId = parseInt(request.user.id);
-      const commentId = parseInt(request.params.commentId);
-      const postId = parseInt(request.params.postId);
-      const post = await Post.findByPk(postId);
-      const commentToDelete = await Comment.findByPk(commentId);
-      if (!commentToDelete) {
-        response.status(404).send({ message: "Comment not found" });
-      } else {
-        if (commentToDelete.userId !== userId) {
-          response
-            .status(403)
-            .send({ message: "You are not authorized to delete this post" });
-        } else {
-          const commentDeleted = await commentToDelete.destroy();
-          await post.update({ comments_num: post.comments_num - 1 });
-        }
-      }
-      return response.status(204).send({ message: "Comment deleted" });
-    } catch (error) {
-      console.log(error);
-    }
-  }
-);
 module.exports = router;
